Refatorado o método de somar.
Refatorado o método de somar alterando os numeradores para decimais e o retorno do método. Assim permite realizar somas com números reais positivos ou reais negativos.

+++ FundamentosTestesUnitarios/Calculadora/Calculadora.cs
@@ -2,7 +2,7 @@
 {
     public class Calculadora
     {
-        public int Somar(int numeroUm, int numeroDois)
+        public decimal Somar(decimal numeroUm, decimal numeroDois)
         {
             return numeroUm + numeroDois;
         }
